Dark Mode and Loading Animations
dark mode is good to go now. some images are designed for light mode, so there are a few sections that don't go dark. ga-giftcard.png was updated to work with dark-mode. also added a simple animated background-img inline svg moving dot for when slideshow images are loading still. figured out how to do it and it does the job.