fix 404 page bug
For some reason, the 404 page wouldn't load the proper custom page I've had on my site ever since switching to a different hosting method. Priority was low on my list. I finally figured out through the command line, by 'get'ing the names of the pages loaded for home and 404, that there were extra characters being added on to the name, causing the file to not be found. The error was related to the name being numbers. Therefore, I changed the name 404 to fourzerofour, directed the website to that upon 404 code. Now it works. yay.